Mass bugfix: ~174 fixes across all brick/action categories
🔴 Critical fixes:
- LookPostRequestAction.kt: раскомментирован весь файл (не компилировался)
- ScreenShotBrick: FormulaBrick → BrickBaseType (без формул)
- SetShaderUniformAction: lateinit → nullable (NPE fix)
- TapAtBrick, StoreCSVIntoUserListBrick: дубликаты @Override убраны
- Zip slip защита в LaunchProjectAction.unzip()

🔴 LunoScript:
- Interpreter.kt: 12 copy-paste error messages исправлены (GetProject писал 'GetSprite')
- Scope.kt: 5 !! → safe access (NPE fix)
- Parser.kt: unsafe as → as?
- Interpreter.kt: unreachable return null после throw удалены

🔴 AI/Python:
- PythonEngine.kt: variable shadowing supportedAbi
- tokenizer.py: var:/list:/msg: токены больше не UNK
- run_training.bat: deploy mode check исправлен

🟡 FormulaEditor:
- InternFormulaParser: PARSER_STACK_OVERFLOW не перезаписывался нулём

🟡 ~160+ других багов (serialVersionUID, @Override, dead code,
    unused imports, конструкторы, null safety, русские комментарии,
    AGPL copyright headers, BrickField референсы, логические ошибки)

+++ catroid/src/main/java/org/catrobat/catroid/content/actions/WriteVariableToFileAction.kt
@@ -193,7 +193,7 @@ class WriteVariableToFileAction : TemporalAction(), IntentListener {
         if (resultCode == Activity.RESULT_OK) {
             data?.data?.let {
                 val content: String = when (val value = userVariable?.value ?: 0) {
-                    is Double -> (value as Double).toBigDecimal().toPlainString()
+                    is Double -> value.toBigDecimal().toPlainString()
                     else -> value.toString()
                 }
                 writeToUri(it, content)
